--- packages/translations-manager/.eslintrc
@@ -1,26 +1,0 @@
-{
-    "extends": [
-        "eslint-config-airbnb-base"
-    ],
-    "env": {
-        "node": true
-    },
-    "rules": {
-        "indent": ["error", 4],
-        "max-len": 0,
-        "no-console": "off",
-        "import/prefer-default-export": 0
-    },
-    "plugins": [
-        "import"
-    ],
-    "parserOptions": {
-        "ecmaVersion": 2018,
-        "sourceType": "module"
-    },
-    "settings": {
-        "import/resolver": {
-            "babel-module": {}
-        }
-    }
-}
