+++ packages/translations-manager/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+    extends: '../../.eslintrc.js',
+    parserOptions: {
+        project: ['./tsconfig.json'],
+    },
+};
