--- packages/translations-manager/babel.config.js
@@ -1,17 +1,0 @@
-module.exports = {
-    presets: [
-        [
-            '@babel/preset-env',
-            {
-                targets: {
-                    node: '8.6.0',
-                },
-            },
-        ],
-    ],
-    plugins: [
-        ['module-resolver', {
-            root: ['./src'],
-        }],
-    ],
-};
