+++ packages/translations-manager/package.json
@@ -1,38 +1,42 @@
 {
-  "name": "@trezor/translations-manager",
-  "version": "1.0.0",
-  "main": "./lib/index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/trezor/trezor-translations-manager.git"
-  },
-  "author": "Trezor <info@trezor.io>",
-  "scripts": {
-    "lint": "eslint .",
-    "clean": "rimraf lib",
-    "build:lib": "yarn run clean && babel src --out-dir lib",
-    "prepublishOnly": "yarn run lint && npm run build"
-  },
-  "bin": {
-    "ttm": "./lib/cli.js"
-  },
-  "license": "SEE LICENSE IN LICENSE.md",
-  "engines": {
-    "node": ">=8.6.0"
-  },
-  "dependencies": {
-    "csv-stringify": "^5.3.0",
-    "csvtojson": "^2.0.8",
-    "dotenv": "^6.2.0",
-    "extract-zip": "^1.6.7",
-    "fs-extra": "^7.0.1",
-    "glob": "^7.1.3",
-    "meow": "^5.0.0",
-    "request": "^2.88.0",
-    "request-promise-native": "^1.0.5"
-  },
-  "devDependencies": {
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-node": "^10.0.0"
-  }
+    "name": "@trezor/translations-manager",
+    "version": "1.0.0",
+    "main": "./lib/index.js",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/trezor/trezor-translations-manager.git"
+    },
+    "author": "Trezor <info@trezor.io>",
+    "scripts": {
+        "lint": "eslint .",
+        "clean": "rimraf lib",
+        "build:lib": "yarn tsc",
+        "prepublishOnly": "yarn run lint && npm run build"
+    },
+    "bin": {
+        "ttm": "./lib/cli.js"
+    },
+    "license": "SEE LICENSE IN LICENSE.md",
+    "engines": {
+        "node": ">=8.6.0"
+    },
+    "dependencies": {
+        "csv-stringify": "^5.3.0",
+        "csvtojson": "^2.0.8",
+        "dotenv": "^6.2.0",
+        "extract-zip": "^1.6.7",
+        "fs-extra": "^7.0.1",
+        "glob": "^7.1.3",
+        "meow": "^5.0.0",
+        "request": "^2.88.0",
+        "request-promise-native": "^1.0.5"
+    },
+    "devDependencies": {
+        "@types/dotenv": "^8.2.0",
+        "@types/extract-zip": "^1.6.2",
+        "@types/meow": "^5.0.0",
+        "@types/request-promise-native": "^1.0.17",
+        "eslint-config-airbnb-base": "^14.0.0",
+        "eslint-plugin-node": "^10.0.0"
+    }
 }
