+++ packages/translations-manager/src/utils/create-row-array.ts
@@ -1,4 +1,12 @@
-const createRowArray = (masterKey, masterValue, occurrences = false) => {
+export interface MessageObject {
+    source: string;
+    meta?: {
+        comment?: string;
+        occurrences?: string[];
+    };
+}
+
+const createRowArray = (masterKey: string, masterValue: MessageObject, occurrences = false) => {
     const context = [];
     if (masterValue.meta) {
         if (masterValue.meta.comment) {
@@ -12,11 +20,7 @@ const createRowArray = (masterKey, masterValue, occurrences = false) => {
         }
     }
 
-    return [
-        masterKey,
-        masterValue.source,
-        context.join(' '),
-    ];
+    return [masterKey, masterValue.source, context.join(' ')];
 };
 
 export default createRowArray;
