+++ packages/translations-manager/tsconfig.json
@@ -1,0 +1,28 @@
+{
+    // "extends": "../../tsconfig.json",
+    "compilerOptions": {
+        "target": "es2017",
+        "module": "CommonJS",
+        "lib": ["es2019", "dom"],
+        "moduleResolution": "node",
+        "allowSyntheticDefaultImports": true,
+        "skipLibCheck": true,
+        "strict": true,
+        "declaration": true,
+        "declarationMap": true,
+        "noImplicitThis": true,
+        "noImplicitAny": true,
+        "strictNullChecks": true,
+        "outDir": "./lib"
+    },
+    "include": ["./src"],
+    "exclude": [
+        "node_modules",
+        "**/node_modules",
+        "**/lib",
+        "**/build",
+        "**/api/generated/**/*.*",
+        "**/_old/**/*.*",
+        ".next"
+    ]
+}
